\name{rbenchmark-package}
\alias{rbenchmark-package}
\docType{package}
\title{\packageTitle{rbenchmark}}
\description{\packageDescription{rbenchmark}}
\section{Package Content}{\packageIndices{rbenchmark}}
\author{\packageAuthor{rbenchmark}}
\section{Maintainer}{\packageMaintainer{rbenchmark}}
\keyword{package}
